alarmpanel: Fix a compiler warning
With GCC 7.1.1 when compiling with -Wall we get

    cc -Wall -O -o alarmpanel alarmpanel.c galaxybus.o -I. -I../AXL -I ../Dataformat ../AXL/axl.o ../Dataformat/dataformat.o -lcurl -lexpat -pthread -lpopt
    alarmpanel.c: In function ‘keypad_update’:
    alarmpanel.c:2011:37: warning: ‘snprintf’ output truncated before the last format character [-Wformat-truncation=]
          snprintf (l2, 17, "ACTIVE %-10s", state_name[t]);
                                     ^
    alarmpanel.c:2011:6: note: ‘snprintf’ output 18 or more bytes into a destination of size 17
          snprintf (l2, 17, "ACTIVE %-10s", state_name[t]);
          ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

This is due to "ACTIVE %-10s" being 18 characters when you include the
'\0', going into device[n].text[1] which is a 17 character buffer.

According to galaxybus.h, the display is 16 characters so the size of
the buffer is correct, so change the format specifier to '%-9s'.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>